requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-RESTful==0.3.10

# Task Queue
redis==5.0.1
celery==5.3.4

# Audio Processing
openai-whisper==20231117
torch>=2.2.0
torchaudio>=2.2.0

# AI/ML
openai==1.3.0
transformers==4.35.0

# Audio/Video Processing
ffmpeg-python==0.2.0
pydub==0.25.1

# Data Processing
pandas==2.1.3
numpy==1.24.3
Pillow==10.0.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring
prometheus-client==0.19.0

# Utilities
requests==2.31.0
click==8.1.7
tqdm==4.66.1 